+++ internal/repository/genre/interface.go
@@ -4,5 +4,5 @@ import "github.com/dangLuan01/rebuild-api-movie28/internal/models"
 
 type GenreRepository interface {
 	FindAll() ([]models.Genre, error)
-	FindBySlug(slug string, page, pageSize int) (models.GenreWithMovie, error)
+	FindBySlug(slug string, page, pageSize int64) (models.GenreWithMovie, error)
 }
